requirements.txt
aiohttp
#aiosocks
aiosqlite
aiohttp_jinja2
requests[socks]
pyyaml
